6-size.c

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	char myChar;
 	float myFloat;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(myChar));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(myInt));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(myLongInt));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(myLongLongInt));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(myFloat));
+	printf("Size of a char: %lu byte(s)\n", sizeof(myChar));
+	printf("Size of an int: %lu byte(s)\n", sizeof(myInt));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(myLongInt));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(myLongLongInt));
+	printf("Size of a float: %lu byte(s)\n", sizeof(myFloat));
 	return (0);
 }
